* fixed reading of retention settings from config file. * startup message now goes through postedMessage instead of being sent directly. * postedMessage now uses values for message retention and QoS from the saved config instead of being hardcoded.

+++ monitor.h
@@ -14,7 +14,7 @@ public:
     class MQTTSettings {
     public:
         QString hostname;
-        unsigned short port;
+        int port;
         QString topCategory;
         QString defaultTopic, willTopic;
         QString startMessage, defaultMessage, willMessage;
